+++ ledger/participant-integration-api/src/main/scala/platform/index/JdbcIndex.scala
@@ -12,7 +12,7 @@ import com.daml.lf.engine.ValueEnricher
 import com.daml.logging.LoggingContext
 import com.daml.metrics.Metrics
 import com.daml.platform.server.api.validation.ErrorFactories
-import com.daml.platform.store.{ReadOnlyLedgerImpl, DbSupport, LfValueTranslationCache}
+import com.daml.platform.store.{DbSupport, LfValueTranslationCache}
 
 import scala.concurrent.ExecutionContext
