+++ ledger/participant-integration-api/src/main/scala/platform/index/LedgerBackedIndexService.scala
@@ -40,7 +40,6 @@ import com.daml.logging.{ContextualizedLogger, LoggingContext}
 import com.daml.platform.ApiOffset
 import com.daml.platform.ApiOffset.ApiOffsetConverter
 import com.daml.platform.server.api.validation.ErrorFactories
-import com.daml.platform.store.ReadOnlyLedger
 import com.daml.platform.store.entries.PartyLedgerEntry
 import com.daml.telemetry.{SpanAttribute, Spans}
 import scalaz.syntax.tag.ToTagOps
