+++ ledger/participant-integration-api/src/main/scala/platform/index/ReadOnlyLedger.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.store
+package com.daml.platform.index
 
 import akka.NotUsed
 import akka.stream.scaladsl.Source
