+++ ledger/participant-integration-api/src/main/scala/platform/store/appendonlydao/events/EventFilter.scala
@@ -1,20 +1,19 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.index
+package com.daml.platform.store.appendonlydao.events
 
 import com.daml.lf.data.Ref
-import com.daml.lf.data.Ref.Party
 import com.daml.ledger.api.domain.TransactionFilter
 import com.daml.ledger.api.v1.event.Event
-import com.daml.ledger.api.v1.value.Identifier
+import com.daml.ledger.api.v1.value.{Identifier => ProtoIdentifier}
 import com.daml.platform.store.Contract.ActiveContract
 import com.daml.platform.api.v1.event.EventOps.EventOps
 
 private[platform] object EventFilter {
 
   // TODO Remove all usages of domain objects
-  private def toLfIdentifier(id: Identifier): Ref.Identifier =
+  private def toLfIdentifier(id: ProtoIdentifier): Ref.Identifier =
     Ref.Identifier(
       packageId = Ref.PackageId.assertFromString(id.packageId),
       qualifiedName = Ref.QualifiedName(
